Reorganize psf utils into common groupings
