+++ photutils/psf/__init__.py
@@ -5,7 +5,7 @@ photometry.
 """
 
 from . import (epsf, epsf_stars, functional_models, gridded_models, groupers,
-               image_models, photometry, utils)
+               image_models, photometry, model_helpers, simulation)
 from .epsf import *  # noqa: F401, F403
 from .epsf_stars import *  # noqa: F401, F403
 from .functional_models import *  # noqa: F401, F403
@@ -14,7 +14,8 @@ from .groupers import *  # noqa: F401, F403
 from .image_models import *  # noqa: F401, F403
 from .matching import *  # noqa: F401, F403
 from .photometry import *  # noqa: F401, F403
-from .utils import *  # noqa: F401, F403
+from .model_helpers import *  # noqa: F401, F403
+from .simulation import *  # noqa: F401, F403
 
 # exclude matching from this list to avoid sphinx warnings
 __all__ = []
@@ -25,4 +26,5 @@ __all__ += gridded_models.__all__
 __all__ += groupers.__all__
 __all__ += image_models.__all__
 __all__ += photometry.__all__
-__all__ += utils.__all__
+__all__ += model_helpers.__all__
+__all__ += simulation.__all__
