+++ photutils/psf/image_models.py
@@ -13,7 +13,7 @@ from astropy.utils.exceptions import AstropyUserWarning
 from photutils.aperture import CircularAperture
 from photutils.utils._parameters import as_pair
 
-__all__ = ['FittableImageModel', 'EPSFModel', 'PRFAdapter']
+__all__ = ['FittableImageModel', 'EPSFModel']
 
 
 class FittableImageModel(Fittable2DModel):
@@ -760,148 +760,3 @@ class EPSFModel(FittableImageModel):
             evaluated_model[invalid] = self._fill_value
 
         return evaluated_model
-
-
-class PRFAdapter(Fittable2DModel):
-    """
-    A model that adapts a supplied PSF model to act as a PRF.
-
-    It integrates the PSF model over pixel "boxes". A critical built-in
-    assumption is that the PSF model scale and location parameters are
-    in *pixel* units.
-
-    Parameters
-    ----------
-    psfmodel : a 2D model
-        The model to assume as representative of the PSF.
-
-    renormalize_psf : bool, optional
-        If True, the model will be integrated from -inf to inf and
-        re-scaled so that the total integrates to 1. Note that this
-        renormalization only occurs *once*, so if the total flux of
-        ``psfmodel`` depends on position, this will *not* be correct.
-
-    flux : float, optional
-        The total flux of the star.
-
-    x_0 : float, optional
-        The x position of the star.
-
-    y_0 : float, optional
-        The y position of the star.
-
-    xname : str or None, optional
-        The name of the ``psfmodel`` parameter that corresponds to the
-        x-axis center of the PSF. If None, the model will be assumed to
-        be centered at x=0.
-
-    yname : str or None, optional
-        The name of the ``psfmodel`` parameter that corresponds to the
-        y-axis center of the PSF. If None, the model will be assumed to
-        be centered at y=0.
-
-    fluxname : str or None, optional
-        The name of the ``psfmodel`` parameter that corresponds to
-        the total flux of the star. If None, a scaling factor will
-        be applied by the ``PRFAdapter`` instead of modifying the
-        ``psfmodel``.
-
-    **kwargs : dict, optional
-        Additional optional keyword arguments to be passed to the
-        `astropy.modeling.Model` parent class.
-
-    Notes
-    -----
-    This current implementation of this class (using numerical
-    integration for each pixel) is extremely slow, and only suited for
-    experimentation over relatively few small regions.
-    """
-
-    flux = Parameter(default=1)
-    x_0 = Parameter(default=0)
-    y_0 = Parameter(default=0)
-
-    def __init__(self, psfmodel, *, renormalize_psf=True, flux=flux.default,
-                 x_0=x_0.default, y_0=y_0.default, xname=None, yname=None,
-                 fluxname=None, **kwargs):
-
-        self.psfmodel = psfmodel.copy()
-
-        if renormalize_psf:
-            from scipy.integrate import dblquad
-            self._psf_scale_factor = 1.0 / dblquad(self.psfmodel,
-                                                   -np.inf, np.inf,
-                                                   lambda x: -np.inf,
-                                                   lambda x: np.inf)[0]
-        else:
-            self._psf_scale_factor = 1
-
-        self.xname = xname
-        self.yname = yname
-        self.fluxname = fluxname
-
-        # these can be used to adjust the integration behavior. Might be
-        # used in the future to expose how the integration happens
-        self._dblquadkwargs = {}
-
-        super().__init__(n_models=1, x_0=x_0, y_0=y_0, flux=flux, **kwargs)
-
-    def evaluate(self, x, y, flux, x_0, y_0):
-        """
-        The evaluation function for PRFAdapter.
-
-        Parameters
-        ----------
-        x, y : float or array_like
-            The coordinates at which to evaluate the model.
-
-        flux : float
-            The total flux of the star.
-
-        x_0, y_0 : float
-            The position of the star.
-
-        Returns
-        -------
-        evaluated_model : `~numpy.ndarray`
-            The evaluated model.
-        """
-        if not np.isscalar(flux):
-            flux = flux[0]
-        if not np.isscalar(x_0):
-            x_0 = x_0[0]
-        if not np.isscalar(y_0):
-            y_0 = y_0[0]
-
-        if self.xname is None:
-            dx = x - x_0
-        else:
-            dx = x
-            setattr(self.psfmodel, self.xname, x_0)
-
-        if self.xname is None:
-            dy = y - y_0
-        else:
-            dy = y
-            setattr(self.psfmodel, self.yname, y_0)
-
-        if self.fluxname is None:
-            return (flux * self._psf_scale_factor
-                    * self._integrated_psfmodel(dx, dy))
-
-        setattr(self.psfmodel, self.yname, flux * self._psf_scale_factor)
-        return self._integrated_psfmodel(dx, dy)
-
-    def _integrated_psfmodel(self, dx, dy):
-        from scipy.integrate import dblquad
-
-        # infer type/shape from the PSF model. Seems wasteful, but the
-        # integration step is a *lot* more expensive so its just peanuts
-        out = np.empty_like(self.psfmodel(dx, dy))
-        outravel = out.ravel()
-        for i, (xi, yi) in enumerate(zip(dx.ravel(), dy.ravel(), strict=True)):
-            outravel[i] = dblquad(self.psfmodel,
-                                  xi - 0.5, xi + 0.5,
-                                  lambda x: yi - 0.5, lambda x: yi + 0.5,
-                                  **self._dblquadkwargs)[0]
-        return out
