+++ photutils/psf/tests/test_image_models.py
@@ -5,10 +5,10 @@ Tests for the image_models module.
 
 import numpy as np
 import pytest
-from astropy.modeling.models import Gaussian2D, Moffat2D
+from astropy.modeling.models import Gaussian2D
 from numpy.testing import assert_allclose
 
-from photutils.psf import FittableImageModel, PRFAdapter
+from photutils.psf import FittableImageModel
 from photutils.utils._optional_deps import HAS_SCIPY
 
 
@@ -127,87 +127,3 @@ class TestFittableImageModel:
         for oversampling in [np.nan, (1, np.inf)]:
             with pytest.raises(ValueError, match=match):
                 FittableImageModel(data, oversampling=oversampling)
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-class TestPRFAdapter:
-    """
-    Tests for PRFAdapter.
-    """
-
-    def normalize_moffat(self, mof):
-        # this is the analytic value needed to get a total flux of 1
-        mof = mof.copy()
-        mof.amplitude = (mof.alpha - 1) / (np.pi * mof.gamma**2)
-        return mof
-
-    @pytest.mark.parametrize('adapterkwargs', [
-        {'xname': 'x_0', 'yname': 'y_0', 'fluxname': None,
-         'renormalize_psf': False},
-        {'xname': None, 'yname': None, 'fluxname': None,
-         'renormalize_psf': False},
-        {'xname': 'x_0', 'yname': 'y_0', 'fluxname': 'amplitude',
-         'renormalize_psf': False}])
-    def test_create_eval_prfadapter(self, adapterkwargs):
-        mof = Moffat2D(gamma=1, alpha=4.8)
-        prf = PRFAdapter(mof, **adapterkwargs)
-
-        # test that these work without errors
-        prf.x_0 = 0.5
-        prf.y_0 = -0.5
-        prf.flux = 1.2
-        prf(0, 0)
-
-    @pytest.mark.parametrize('adapterkwargs', [
-        {'xname': 'x_0', 'yname': 'y_0', 'fluxname': None,
-         'renormalize_psf': True},
-        {'xname': 'x_0', 'yname': 'y_0', 'fluxname': None,
-         'renormalize_psf': False},
-        {'xname': None, 'yname': None, 'fluxname': None,
-         'renormalize_psf': False}])
-    def test_prfadapter_integrates(self, adapterkwargs):
-        from scipy.integrate import dblquad
-
-        mof = Moffat2D(gamma=1.5, alpha=4.8)
-        if not adapterkwargs['renormalize_psf']:
-            mof = self.normalize_moffat(mof)
-        prf1 = PRFAdapter(mof, **adapterkwargs)
-
-        # first check that the PRF over a central grid ends up summing to the
-        # integrand over the whole PSF
-        xg, yg = np.meshgrid(*([(-1, 0, 1)] * 2))
-        evalmod = prf1(xg, yg)
-
-        if adapterkwargs['renormalize_psf']:
-            mof = self.normalize_moffat(mof)
-
-        integrand, itol = dblquad(mof, -1.5, 1.5, lambda x: -1.5,
-                                  lambda x: 1.5)
-        assert_allclose(np.sum(evalmod), integrand, atol=itol * 10)
-
-    @pytest.mark.parametrize('adapterkwargs', [
-        {'xname': 'x_0', 'yname': 'y_0', 'fluxname': None,
-         'renormalize_psf': False},
-        {'xname': None, 'yname': None, 'fluxname': None,
-         'renormalize_psf': False}])
-    def test_prfadapter_sizematch(self, adapterkwargs):
-        from scipy.integrate import dblquad
-
-        mof1 = self.normalize_moffat(Moffat2D(gamma=1, alpha=4.8))
-        prf1 = PRFAdapter(mof1, **adapterkwargs)
-
-        # now try integrating over differently-sampled PRFs
-        # and check that they match
-        mof2 = self.normalize_moffat(Moffat2D(gamma=2, alpha=4.8))
-        prf2 = PRFAdapter(mof2, **adapterkwargs)
-
-        xg1, yg1 = np.meshgrid(*([(-0.5, 0.5)] * 2))
-        xg2, yg2 = np.meshgrid(*([(-1.5, -0.5, 0.5, 1.5)] * 2))
-
-        eval11 = prf1(xg1, yg1)
-        eval22 = prf2(xg2, yg2)
-
-        _, itol = dblquad(mof1, -2, 2, lambda x: -2, lambda x: 2)
-        # it's a bit of a guess that the above itol is appropriate, but
-        # it should be close
-        assert_allclose(np.sum(eval11), np.sum(eval22), atol=itol * 100)
